services/remediation/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain-openai
langchain-core
PyGithub
